Move shared utils to their own directory
