+++ contentcuration/contentcuration/frontend/shared/data/changes.js
@@ -2,7 +2,7 @@ import Dexie from 'dexie';
 import uuidv4 from 'uuid/v4';
 import { EventEmitter } from 'events';
 import db, { CLIENTID } from 'shared/data/db';
-import { promiseChunk } from 'shared/utils';
+import { promiseChunk } from 'shared/utils/helpers';
 import {
   CHANGE_LOCKS_TABLE,
   CHANGES_TABLE,
