+++ contentcuration/contentcuration/frontend/shared/utils/helpers.js
